fix libretrobridge to be 64bit only (and fix it for 64bit)

+++ BizHawk.Emulation.Cores/Libretro/LibretroApi_QUERY.cs
@@ -6,11 +6,11 @@ namespace BizHawk.Emulation.Cores.Libretro
 {
 	unsafe partial class LibretroApi
 	{
-		public Tuple<IntPtr, int> QUERY_GetMemory(RETRO_MEMORY mem)
+		public Tuple<IntPtr, ulong> QUERY_GetMemory(RETRO_MEMORY mem)
 		{
 			comm->value = (uint)mem;
 			Message(eMessage.QUERY_GetMemory);
-			return Tuple.Create(new IntPtr(comm->buf[(int)BufId.Param0]), comm->buf_size[(int)BufId.Param0]);
+			return Tuple.Create(new IntPtr((long)comm->buf[(int)BufId.Param0]), comm->buf_size[(int)BufId.Param0]);
 		}
 	}
 }
